+++ chunk.h
@@ -84,7 +84,7 @@ typedef struct {
 
 void initChunk(Chunk* chunk);
 void freeChunk(Chunk* chunk);
-void writeChunk(Chunk* chunk, uint8_t byte, int line);
+void writeChunk(Chunk* chunk, int byte, int line);
 int  addConstant(Chunk* chunk, Value value);
 int  getLine(Chunk* chunk, int offset);
